device: don't wait udev for devices created by us
If we created a software interface it is because we already decided
that it should be managed, and so there is no point in waiting udev to
check that the interface is not udev-unmanaged.

We still wait udev for software interfaces created externally.

(cherry picked from commit 58db61545d101723e7cb45dc7eadc7e4ad7a4b0c)