dispatcher: don't print debug messages of dispatcher in regular mode
Previously, we would log several messages with level "debug" / g_info().

  _LOG_R_D (request, "start running ordered scripts...");
  _LOG_R_D (request, "new request (%u scripts)", request->scripts->len);
  _LOG_R_D (request, "completed: no scripts");

Note that this effectively logs a message for every event. I think that
is to verbose and not suitable for regular logging.

Only enable these messages if debug logging is enabled. As such, these debug
level messsages now are enabled together with the trace level messages.